Update lần 6 by Quoc Cuong

+++ database/migrations/2017_12_06_150735_create_sanpham_table.php
@@ -19,8 +19,8 @@ class CreateSanphamTable extends Migration
             $table->integer('idloai');
             $table->string('description',255);
             $table->string('image',100);
-            $table->float('gia');
-            $table->float('giakm');
+            $table->int('gia');
+            $table->int('giakm');
             $table->integer('soluong');
             $table->integer('new');
             $table->timestamps();
